+++ allink_core/apps/people/sitemaps.py
@@ -1,12 +1,12 @@
 # -*- coding: utf-8 -*-
-from django.contrib.sitemaps import Sitemap
 
 from allink_core.core.loading import get_model
+from allink_core.core.sitemap import HrefLangSitemap
 
 People = get_model('people', 'People')
 
 
-class PeopleSitemap(Sitemap):
+class PeopleSitemap(HrefLangSitemap):
 
     changefreq = "never"
     priority = 0.5